% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data_format.R
\name{check_data_format}
\alias{check_data_format}
\title{Check forecast data structure}
\usage{
check_data_format(forecast_data)
}
\arguments{
\item{forecast_data}{A data.frame with the predictions that is or can be coerced to a model_out_tbl format.}
}
\description{
Check forecast data structure
}
