% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_data_validation.R
\name{input_data_validation}
\alias{input_data_validation}
\title{Check the input forecast data structure}
\usage{
input_data_validation(forecast_data)
}
\arguments{
\item{forecast_data}{A data.frame with the predictions that is or can be coerced to a model_out_tbl format.}
}
\value{
a model_out_tbl format that has a single output type
}
\description{
Check the input forecast data structure
}
\examples{
library(dplyr)
library(hubUtils)
hub_path <- system.file("testhubs/flusight", package = "hubUtils")
hub_con <- connect_hub(hub_path)
hub_con \%>\%
        filter(output_type == "quantile") \%>\%
        collect() \%>\%
        input_data_validation()
}
