% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble_component_importance_score.R
\name{ensemble_component_importance_score}
\alias{ensemble_component_importance_score}
\title{Evaluate ensemble component model's importance based on a measure of their contribution to ensemble prediction accuracy}
\usage{
ensemble_component_importance_score(
  forecast_data,
  true_value,
  output_type = c("quantile", "pmf", "cdf", "mean", "median", "sample"),
  agg_fun = c("mean", "median"),
  weighted = FALSE,
  method = c("lomo", "lasomo"),
  subset_wt = c("equal", "perm_based"),
  scoring_rule = c("MAE", "MSE", "WIS", "CRPS", "Logscore"),
  na.action = c("worst", "average", "drop")
)
}
\arguments{
\item{forecast_data}{A data.frame with the predictions that is or can be coerced to a model_out_tbl format.}

\item{true_value}{Ground truth data, i.e., "target data", for the variables that are used to define modeling targets.}

\item{output_type}{String specifying the type of representation of the predictive distribution;  \code{c("quantile", "pmf", "cdf", "mean", "median", "sample")}.}

\item{agg_fun}{String specifying method for ensembling quantiles; \code{c("mean", "median")}.}

\item{weighted}{Boolean indicating whether model weighting should be done. If \code{FALSE}, all models are given equal weight. If \code{TRUE}, model weights are estimated.}

\item{method}{String specifying algorithm for model importance calculation; \code{c("lomo", "lasomo")}. \code{"lomo"} stands for leave-one-model-out and \code{"lasomo"} stands for leave all subsets of models out.}

\item{subset_wt}{String specifying method for assigning weight to subsets when using \code{lasomo} algorithm.}

\item{scoring_rule}{String specifying metric to use to calculate importance; \code{c("MAE", "MSE", "WIS", "CRPS", "Logscore")}.}

\item{na.action}{String specifying treatment for missing data; \code{c("worst", "average", "drop")}.}
}
\value{
A data.frame with columns \code{task_id}, \code{output_type}, \code{model}, \code{importance_score}.
}
\description{
Evaluate ensemble component model's importance based on a measure of their contribution to ensemble prediction accuracy
}
